- Added a property sheet Global.props which will contain settings that should be globally applied to all projects in all configurations.  In this case, it's used for setting the warning level /W4 on everyone. - All projects in both the msvc and c::b builds now have warnings set as high as I want them /W4 (msvc) or -Wall (gcc). - Addressed all warnings in both the msvc and gcc builds. - Most of the warnings were just unreferenced formal parameters and signed/unsigned comparisons, but a few were substantial: - UserInterface::DisplayPrompt was using the wrong value in its if-checks (result instead of windowsResult). refs #109
- Added PlatformDefines.h which detects the compiler, platform, and architecture being built.  This is useful for detecting msvc/gcc to determine whether or not the custom secure crt re-implementation is necessary.
- Added SecureCrt.h.  This is only used in non-msvc builds.  Its goal is to re-implement Microsoft's secure crt functions as needed.  Right now, it only properly implements one or two, and the rest are just insecure wrappers.
refs #110

+++ Source/WindowsApplication/WaveOutSound/WaveOutSound.cpp
@@ -75,7 +75,7 @@ public:
 
 		_Mute = false;
 
-		for(int i=0;i<AudioBuffer::BufferSizeSamples;i++)
+		for(unsigned int i=0;i<AudioBuffer::BufferSizeSamples;i++)
 		{
 			_Silence.Samples[0][i] = SilentSample;
 			_Silence.Samples[1][i] = SilentSample;
@@ -136,7 +136,7 @@ public:
 
 		while(_Phoenix->ShutdownRequested() == false)
 		{
-			if(_Mute == false && _Machine != NULL && _Machine->GetSound()->GetAudioBufferCount() != _LastFrameQueued)
+			if(_Mute == false && _Machine != NULL && _Machine->GetSound()->GetAudioBufferCount() != (int)_LastFrameQueued)
 			{
 				AudioBuffer buffer = _Machine->GetSound()->GetStableAudioBuffer();
